+++ src/commands/mod.rs
@@ -1,5 +1,3 @@
 pub mod ping;
 pub mod getters;
 pub mod setters;
-pub mod joke;
-pub mod blacklist;
