+++ src/commands/setters/set_timeout_timer.rs
@@ -1,7 +1,7 @@
 use crate::DB;
 use crate::utils::{CommandResult, Context};
-use crate::utils::autocomplete::args_set_timeout_timer;
-use crate::utils::debug::UnwrapLog;
+use crate::utils::misc::autocomplete::args_set_timeout_timer;
+use crate::utils::misc::debug::UnwrapLog;
 use crate::commands::setters::SetTimeoutTimer;
 
 /// Establece el tiempo de timeout para el servidor
