+++ src/commands/setters/set_warn_message.rs
@@ -1,6 +1,6 @@
 use crate::DB;
 use crate::utils::{CommandResult, Context};
-use crate::utils::debug::UnwrapLog;
+use crate::utils::misc::debug::UnwrapLog;
 use crate::commands::setters::WarnMessageData;
 
 /// Establece el mensaje de advertencia si se menciona a un usuario o un usuario con un rol prohíbido
