+++ src/utils/handlers/edited_messages.rs
@@ -3,9 +3,9 @@ use poise::serenity_prelude as serenity;
 use crate::commands::setters::ForbiddenRoleData;
 use crate::commands::setters::ForbiddenUserData;
 use crate::commands::setters::GuildData;
-use crate::utils::debug::UnwrapLog;
+use crate::utils::misc::debug::UnwrapLog;
 use crate::utils::Error;
-use crate::utils::embeds::{edit_message_embed, edit_message_embed_if_mention};
+use crate::utils::misc::embeds::{edit_message_embed, edit_message_embed_if_mention};
 use crate::utils::handlers::misc::forbidden_mentions::{handle_forbidden_user, handle_forbidden_role};
 use crate::utils::MessageData;
 use crate::matches;
