+++ src/utils/misc/embeds.rs
