+++ src/utils/misc/mod.rs
@@ -1,0 +1,3 @@
+pub mod debug;
+pub mod embeds;
+pub mod autocomplete;
